Change all uint32_t to size_t. Make sure things are const when they should be const.

+++ test/test_read_data/read_data.cpp
@@ -12,7 +12,7 @@ int main(int argc, char* argv[]) {
     data_io.read_data_from_file(rockstar_data);
 
     // check the types of all of the keys
-    uint32_t total_keys = rockstar_data.get_total_keys();
+    size_t total_keys = rockstar_data.get_total_keys();
     for (auto i = 0; i < total_keys; i++) {
         std::cout << "Index: " << i << " Key: " << rockstar_data.get_key(i) << std::endl;
     }
@@ -23,17 +23,17 @@ int main(int argc, char* argv[]) {
     std::cout << "mvir at 0 is " << std::get<real>(rockstar_data.data_.at(0)->at(rockstar_data.get_key("virial_mass"))) << std::endl;
 
     real mvir;
-    uint32_t mvir_key = rockstar_data.get_key("virial_mass");
+    size_t mvir_key = rockstar_data.get_key("virial_mass");
     // print the first 10 mvir values
-    for (uint32_t i = 0; i < 10; i++) {
+    for (size_t i = 0; i < 10; i++) {
         rockstar_data.data_at(mvir, i, mvir_key);
         std::cout << "mvir #" << i << " = " << mvir << " Msun" << std::endl;
     }
 
     int64_t id;
-    uint32_t id_key = rockstar_data.get_key("id");
+    size_t id_key = rockstar_data.get_key("id");
     // print the first 10 id values
-    for (uint32_t i = 0; i < 10; i++) {
+    for (size_t i = 0; i < 10; i++) {
         rockstar_data.data_at(id, i, id_key);
         std::cout << "id #" << i << " = " << id << std::endl;
     }
